Cleaned up most of the remaining GCC warnings.
SVN r1459 (latestzdoom)

+++ src/browser.cpp
@@ -920,12 +920,12 @@ CCMD( dumpserverlist )
 		if ( g_BrowserServerList[ulIdx].ulActiveState != AS_ACTIVE )
 			continue;
 
-		Printf( "\nServer #%d\n----------------\n", ulIdx );
+		Printf( "\nServer #%d\n----------------\n", static_cast<unsigned int> (ulIdx) );
 		Printf( "Name: %s\n", g_BrowserServerList[ulIdx].HostName.GetChars() );
 		Printf( "Address: %s\n", NETWORK_AddressToString( g_BrowserServerList[ulIdx].Address ));
 		Printf( "Gametype: %d\n", g_BrowserServerList[ulIdx].GameMode );
-		Printf( "Num PWADs: %d\n", g_BrowserServerList[ulIdx].lNumPWADs );
-		Printf( "Players: %d/%d\n", g_BrowserServerList[ulIdx].lNumPlayers, g_BrowserServerList[ulIdx].lMaxClients );
-		Printf( "Ping: %d\n", g_BrowserServerList[ulIdx].lPing );
+		Printf( "Num PWADs: %d\n", static_cast<int> (g_BrowserServerList[ulIdx].lNumPWADs) );
+		Printf( "Players: %d/%d\n", static_cast<int> (g_BrowserServerList[ulIdx].lNumPlayers), static_cast<int> (g_BrowserServerList[ulIdx].lMaxClients) );
+		Printf( "Ping: %d\n", static_cast<int> (g_BrowserServerList[ulIdx].lPing) );
 	}
 }
